requirements.txt
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-talisman==0.8.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.19.1
six==1.16.0
Werkzeug==2.0.2
requests==2.26.0
firebase_admin==5.2.0
ratelimit==2.2.1
cachetools==5.2.0
cryptography==35.0.0
slack_sdk==3.18.1
markdown==3.4.1
mock-firestore==0.11.0
ratelimiter==1.2.0
pygithub==2.1.1
openai==1.10.0
asgiref==3.7.2
GitPython==3.1.41
typing==3.7.4.3
dataclasses==0.6
git-fame==2.0.1
qrcode==7.4.2
pillow==10.2.0
pytz==2023.4
propelauth_flask==2.1.11
littletable==2.2.4
et-xmlfile==1.1.0
openpyxl==3.1.2